chore: clean up font and move to shared folder
